fix: Update InstanceV1/V2 related types based on real-world usage
Many servers that claim to be Mastodon-API compatible are not, as
evidenced by the content they include in the responses to
`/api/v1/instance` and `/api/v2/instance` requests.

Work around the worst of the breakage by providing defaults or marking
some fields as nullable (with a default null).

Bugs have been reported against the relevant projects.

+++ app/src/main/java/app/pachli/components/compose/ImageDownsizer.kt
@@ -36,7 +36,7 @@ import java.io.FileOutputStream
  */
 fun downsizeImage(
     uri: Uri,
-    sizeLimit: Int,
+    sizeLimit: Long,
     contentResolver: ContentResolver,
     tempFile: File,
 ): Boolean {
